# lib/api-zod/src/generated/types/candidate.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CandidateGender } from "./candidateGender";

export interface Candidate {
  id: number;
  firstName: string;
  lastName: string;
  gender: CandidateGender;
  dateOfBirth: string;
  email: string;
  phoneNumber: string;
  courseOfStudy: string;
  percentageScored: number;
  nameOfCollege: string;
  hasBacklog: boolean;
  createdAt: string;
}