# lib/api-zod/src/generated/types/createCandidate.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CreateCandidateGender } from "./createCandidateGender";

export interface CreateCandidate {
  /** @minLength 1 */
  firstName: string;
  /** @minLength 1 */
  lastName: string;
  gender: CreateCandidateGender;
  /** Date of birth in YYYY-MM-DD format */
  dateOfBirth: string;
  email: string;
  /** @minLength 1 */
  phoneNumber: string;
  /** @minLength 1 */
  courseOfStudy: string;
  /**
   * @minimum 0
   * @maximum 100
   */
  percentageScored: number;
  /** @minLength 1 */
  nameOfCollege: string;
  hasBacklog: boolean;
}