# lib/db/src/schema/candidates.ts
import { pgTable, serial, text, numeric, boolean, date, timestamp } from "drizzle-orm/pg-core";
import { createInsertSchema } from "drizzle-zod";
import { z } from "zod/v4";

export const candidatesTable = pgTable("candidates", {
  id: serial("id").primaryKey(),
  firstName: text("first_name").notNull(),
  lastName: text("last_name").notNull(),
  gender: text("gender").notNull(),
  dateOfBirth: date("date_of_birth").notNull(),
  email: text("email").notNull(),
  phoneNumber: text("phone_number").notNull(),
  courseOfStudy: text("course_of_study").notNull(),
  percentageScored: numeric("percentage_scored", { precision: 5, scale: 2 }).notNull(),
  nameOfCollege: text("name_of_college").notNull(),
  hasBacklog: boolean("has_backlog").notNull(),
  createdAt: timestamp("created_at").defaultNow().notNull(),
});

export const insertCandidateSchema = createInsertSchema(candidatesTable).omit({
  id: true,
  createdAt: true,
});

export type InsertCandidate = z.infer<typeof insertCandidateSchema>;
export type Candidate = typeof candidatesTable.$inferSelect;